+++ enterprise/frontend/src/embedding-sdk/components/public/SdkQuestion/SdkQuestion.stories.tsx
@@ -8,7 +8,7 @@ import {
 } from "embedding-sdk/test/storybook-id-args";
 import { Box } from "metabase/ui";
 
-import { SdkQuestion } from "./InteractiveQuestion";
+import { SdkQuestion } from "./SdkQuestion";
 
 const QUESTION_ID = (window as any).QUESTION_ID || questionIds.numberId;
