+++ enterprise/frontend/src/embedding-sdk/components/public/SdkQuestion/SdkQuestion.tsx
