+++ enterprise/frontend/src/embedding-sdk/components/public/SdkQuestion/SdkQuestion.unit.spec.tsx
@@ -43,10 +43,7 @@ import { createMockEntityId } from "metabase-types/api/mocks/entity-id";
 
 import { useSdkQuestionContext } from "../../private/SdkQuestion/context";
 
-import {
-  type BaseInteractiveQuestionProps,
-  SdkQuestion,
-} from "./InteractiveQuestion";
+import { type BaseInteractiveQuestionProps, SdkQuestion } from "./SdkQuestion";
 const TEST_PARAM = createMockParameter({
   type: "number/=",
   slug: "product_id",
