+++ enterprise/frontend/src/embedding-sdk/components/public/SdkQuestion/index.ts
@@ -24,4 +24,4 @@ export {
   type BaseInteractiveQuestionProps,
   type InteractiveQuestionProps,
   type DrillThroughQuestionProps,
-} from "./InteractiveQuestion";
+} from "./SdkQuestion";
